Move field components
